+++ packages/ub/CHANGELOG.md
@@ -12,7 +12,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - automatically creation of BLOB store structure - no need to call `ubcli createStore` anymore. 
   In case of DFS folders should be created/mounted manually
   
-
 ## [4.0.30]
 ### Added
 
@@ -20,4 +19,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `mdb` virtual data store correctly handle models with public path only 
 - `clientRequire` endpoint return correct mime type for files (using mime-db)
 - optimize `clientRequire` endpoint by caching resolved path's to globalCache
+- UB.UBAbort server side exception stack trace now independent of UB.js placement